credit_risk_ml_system/requirements.txt
pandas
numpy
scikit-learn
mlflow
shap
xgboost
flask
joblib
scipy
pyyaml